REFACT: organize the classes (commands and controllers) related to auth in an auth module / namespace
