--- app/controllers/api/v1/sign_out_controller.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    class SignOutController < BaseController
-      def destroy
-        Auth::SignOutCommand.call(params: { token: current_session.token })
-
-        head :ok
-      end
-    end
-  end
-end
